Extract shell heredoc symbol partial
